fix(server): remove circular dependency
